pyproject.toml
[tool.poetry]
name = "website_availability_monitor_kafka"
version = "0.1.0"
description = ""
authors = ["kbaikov <kbaikov@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
psycopg2 = "^2.8.6"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
mypy = "^0.790"
pre-commit = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
